Patch i18n path lookup function
Adds an additional sed command that patches a second function the
desktop code uses to look up its i18n files. It's worth noting that the
path the original code looks for is in the `i18n` subdirectory, but the
files themselves wind up in the normal resources directory. Since it
doesn't really matter where they wind up as long as we can find them, I
didn't dig into that, but it seems worth noting in case something
changes in the build down the line.

Fixes #1.